image_and_video/src/image_and_video/requirements.txt
streamlit>=1.22.0
Pillow>=9.0.0
boto3>=1.38.18
numpy>=1.20.0
requests>=2.28.0
python-resize-image>=1.1.20
streamlit-cognito-auth
Authlib>=1.3.2
python-dotenv
pyjwt